modify : checkSpaceWallTemporary메소드 SpaceWallTemporaryService로 이동 (#114)
